// src/models/IOnlineAgenda.ts
export interface IOnlineAgenda {
  surname: string
  photoUrl: string
  name: string
  businessName: string
  eventDay: number
  state: string
  startDate: string
  endDate: string
  startHour: string
  endHour: string
  buyerId: number
  stateId: number
  meetingId: number
  sellerId: number
  originator: string
  selectable: boolean
  stand: string
}