// src/models/IOriginalOnlineAgenda.ts
export interface IOriginalOnlineAgenda {
  ContraparteApellido: string
  ContraparteFotoFile: string
  ContraparteNombre: string
  ContraparteRazonSocial: string
  DiaEvento: number
  Estado: string
  FechaFin: string
  FechaInicio: string
  HoraFin: string
  HoraInicio: string
  IDComprador: number
  IDEstado: number
  IDReunion: number
  IDVendedor: number
  Originante: string
  Seleccionable: boolean
  Stand: string
}